Move level modification commands into LevelsCommand
This commit creates a new slash command that is used to set, get, and view the top levels.
